Move general-purpose diff/editscript code to common.diff package
- Move ShortestEditScript.kt to common.diff package
- Extract TokenCursor to EditScriptCursor in common.diff package, make public
- Move ShortestEditScriptTest to common.diff package
- Update imports in EditLetterDiff and EditScriptTest
