+++ skribenten-backend/src/main/kotlin/no/nav/pensjon/brev/skribenten/letter/EditLetterDiff.kt
@@ -5,6 +5,10 @@ package no.nav.pensjon.brev.skribenten.letter
 import no.nav.brev.InterneDataklasser
 import no.nav.brev.Listetype
 import no.nav.pensjon.brev.skribenten.common.EqualityBy
+import no.nav.pensjon.brev.skribenten.common.diff.EditOperation
+import no.nav.pensjon.brev.skribenten.common.diff.EditScript
+import no.nav.pensjon.brev.skribenten.common.diff.EditScriptCursor
+import no.nav.pensjon.brev.skribenten.common.diff.shortestEditScript
 import no.nav.pensjon.brev.skribenten.letter.ContentIndex.BlockContentIndex
 import no.nav.pensjon.brev.skribenten.letter.ContentIndex.ItemContentIndex
 import no.nav.pensjon.brev.skribenten.letter.ContentIndex.TableCellContentIndex
@@ -93,7 +97,7 @@ class EditLetterWordDiff : EditLetterDiff<EditLetterWordDiff.Token> {
     )
 
     private fun generateDiffSegments(tokens: List<Token>, edits: List<EditOperation<Token>>): List<DiffSegment> {
-        val cursor = TokenCursor(tokens, edits)
+        val cursor = EditScriptCursor(tokens, edits)
         return cursor.flatMapIndexed<Token.Block, DiffSegment> { blockIndex, _, _ ->
             BlockParser(blockIndex, cursor).parse()
         }.also { require(!cursor.hasNext) { "Not all tokens were consumed, next: ${cursor.peek()}" } }
@@ -130,7 +134,7 @@ class EditLetterWordDiff : EditLetterDiff<EditLetterWordDiff.Token> {
     // This is a general block parser which will not fail for Title-blocks that contain invalid content types.
     private class BlockParser(
         private val blockIndex: Int,
-        private val cursor: TokenCursor<Token>,
+        private val cursor: EditScriptCursor<Token>,
     ) {
         fun parse(): List<DiffSegment> =
             cursor.flatMapIndexed<Token.BlockContent, DiffSegment> { currentPosition, token, _ ->
@@ -196,54 +200,5 @@ class EditLetterWordDiff : EditLetterDiff<EditLetterWordDiff.Token> {
                     is Token.NewLine -> emptyList()
                 }
             }
-    }
-}
-
-private class TokenCursor<T : Any>(private val tokens: List<T>, edits: List<EditOperation<T>>) {
-    init {
-        require(edits.distinctBy { it.position }.size == edits.size) { "Expected edits to only have unique position references" }
-    }
-    private val edits = edits.associateBy { it.position }
-    private var currentIndex = 0
-    val hasNext: Boolean get() = currentIndex < tokens.size
-
-    fun peek(): T? = tokens.getOrNull(currentIndex)
-
-    fun consume(): Pair<T, EditOperation<T>?> = Pair(tokens[currentIndex], edits[currentIndex++]).also {
-        require(it.second == null || it.second?.value == it.first) {
-            "Expected edit operation value to match tokens at position ${currentIndex - 1}, but was ${it.second?.value} and ${it.first}"
-        }
-    }
-
-    inline fun <reified E : T> requireAndConsume(): E {
-        val (token) = consume()
-        require(token is E) { "Expected to consume ${E::class.simpleName}-token but found: $token" }
-        return token
-    }
-
-    inline fun <reified E : T> consumeIf(): Pair<E, EditOperation<E>?>? {
-        if (peek() !is E) return null
-        val (token, edit) = consume()
-        @Suppress("UNCHECKED_CAST")
-        return Pair(token as E, edit as EditOperation<E>?)
-    }
-
-    inline fun <reified E : T, R> fold(initial: R, action: (R, E, EditOperation<E>?) -> R): R {
-        var accumulator = initial
-        while (true) {
-            val (token, edit) = consumeIf<E>() ?: break
-            accumulator = action(accumulator, token, edit)
-        }
-        return accumulator
-    }
-
-    inline fun <reified E : T, R> flatMapIndexed(action: (Int, E, EditOperation<E>?) -> List<R>): List<R> {
-        val result = mutableListOf<R>()
-        var index = 0
-        while (true) {
-            val (token, edit) = consumeIf<E>() ?: break
-            result.addAll(action(index++, token, edit))
-        }
-        return result
     }
 }
